# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fable Vehicle Recognition Backend",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/detections": {
      "get": {
        "tags": [
          "detections"
        ],
        "summary": "List Detections",
        "operationId": "list_detections_api_v1_detections_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vehicle Type"
            }
          },
          {
            "name": "manufacturer",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Manufacturer"
            }
          },
          {
            "name": "license_plate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "License Plate"
            }
          },
          {
            "name": "image_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Image Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedDetections"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/detections/{detection_id}/plate-crop": {
      "get": {
        "tags": [
          "detections"
        ],
        "summary": "Get Plate Crop",
        "description": "Serve the plate crop kept as review evidence.\n\nBoth variants exist so a reviewer can judge the geometry stage, not just its output:\n'after' is missing when the corners could not be recovered and no warp was applied.",
        "operationId": "get_plate_crop_api_v1_detections__detection_id__plate_crop_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "detection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Detection Id"
            }
          },
          {
            "name": "variant",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "before",
                "after"
              ],
              "type": "string",
              "description": "'before' is the detector crop as found; 'after' is the same crop with the perspective removed.",
              "default": "after",
              "title": "Variant"
            },
            "description": "'before' is the detector crop as found; 'after' is the same crop with the perspective removed."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/detections/{detection_id}": {
      "get": {
        "tags": [
          "detections"
        ],
        "summary": "Get Detection",
        "operationId": "get_detection_api_v1_detections__detection_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "detection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Detection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "detections"
        ],
        "summary": "Patch Detection",
        "operationId": "patch_detection_api_v1_detections__detection_id__patch",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "detection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Detection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectionPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "detections"
        ],
        "summary": "Delete Detection",
        "operationId": "delete_detection_api_v1_detections__detection_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "detection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Detection Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BBoxOut": {
        "properties": {
          "x1": {
            "type": "number",
            "title": "X1"
          },
          "y1": {
            "type": "number",
            "title": "Y1"
          },
          "x2": {
            "type": "number",
            "title": "X2"
          },
          "y2": {
            "type": "number",
            "title": "Y2"
          }
        },
        "type": "object",
        "required": [
          "x1",
          "y1",
          "x2",
          "y2"
        ],
        "title": "BBoxOut"
      },
...
      "DetectionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "image_id": {
            "type": "string",
            "format": "uuid",
            "title": "Image Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "vehicle_type": {
            "type": "string",
            "title": "Vehicle Type"
          },
          "manufacturer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "license_plate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Plate"
          },
          "plate_readable": {
            "type": "boolean",
            "title": "Plate Readable"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "bbox": {
            "$ref": "#/components/schemas/BBoxOut"
          },
          "extra": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "vehicle_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Confidence"
          },
          "manufacturer_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer Confidence"
          },
          "source_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Label"
          },
          "vehicle_type_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Type Source"
          },
          "vehicle_type_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Type Confidence"
          },
          "plate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlateOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_id",
          "job_id",
          "vehicle_type",
          "manufacturer",
          "model",
          "license_plate",
          "plate_readable",
          "confidence",
          "bbox",
          "extra",
          "created_at",
          "updated_at"
        ],
        "title": "DetectionOut"
      },
      "DetectionPatch": {
        "properties": {
          "vehicle_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Type"
          },
          "manufacturer": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "license_plate": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "License Plate"
          },
          "plate_readable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plate Readable"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NormalizedBBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "plate_bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NormalizedBBox"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DetectionPatch"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NormalizedBBox": {
        "properties": {
          "x1": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X1"
          },
          "y1": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y1"
          },
          "x2": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X2"
          },
          "y2": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y2"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x1",
          "y1",
          "x2",
          "y2"
        ],
        "title": "NormalizedBBox"
      },
      "PaginatedDetections": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DetectionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedDetections"
      },
      "PlateOut": {
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BBoxOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "quad": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quad"
          },
          "detection_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detection Confidence"
          },
          "ocr_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Confidence"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "text_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Raw"
          },
          "readable": {
            "type": "boolean",
            "title": "Readable"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "ocr_engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Engine"
          },
          "ocr_variant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Variant"
          },
          "rectified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rectified"
          },
          "perspective_skew": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perspective Skew"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "crop_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop Filename"
          },
          "rectified_crop_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rectified Crop Filename"
          }
        },
        "type": "object",
        "required": [
          "bbox",
          "quad",
          "detection_confidence",
          "ocr_confidence",
          "text",
          "text_raw",
          "readable",
          "format",
          "region",
          "ocr_engine",
          "ocr_variant",
          "rectified",
          "perspective_skew",
          "status",
          "crop_filename",
          "rectified_crop_filename"
        ],
        "title": "PlateOut",
        "description": "Plate geometry and reading, reported separately from the vehicle."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}